feat: only original lender can update/delete loan
Created a policy that will verify that only original lender can update and delete the loan.

+++ app/Http/Controllers/API/V1/LoanController.php
@@ -7,8 +7,9 @@ use App\Http\Requests\API\V1\StoreLoanRequest;
 use App\Http\Requests\API\V1\UpdateLoanRequest;
 use App\Http\Resources\V1\LoanResource;
 use App\Models\Loan;
+use Exception;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class LoanController extends Controller
 {
@@ -53,6 +54,11 @@ class LoanController extends Controller
         try {
             $loan = Loan::findOrFail($id);
 
+            // Check if the user is the lender of the loan being updated
+            if ($request->user()->cannot('update', $loan)) {
+                throw new Exception('Access denied: only the lender can update this loan.', 403);
+            }
+
             $loan->update($request->mappedAttributes());
 
             return new LoanResource($loan);
@@ -61,6 +67,11 @@ class LoanController extends Controller
                 'message' => 'Loan cannot be found.',
                 'status' => 404
             ], 404);
+        } catch (Exception $e) {
+            return response()->json([
+                'message' => $e->getMessage(),
+                'status' => $e->getCode()
+            ], $e->getCode());
         }
     }
 
@@ -71,6 +82,12 @@ class LoanController extends Controller
     {
         try {
             $loan = Loan::findOrfail($id);
+
+            // Check if the user is the lender of the loan being updated
+            if (Auth::user()->cannot('delete', $loan)) {
+                throw new Exception('Access denied: only the lender can delete this loan.', 403);
+            }
+
             $loan->delete();
 
             return response()->json([
@@ -82,6 +99,11 @@ class LoanController extends Controller
                 'message' => 'Loan cannot be found.',
                 'status' => 404
             ], 404);
+        } catch (Exception $e) {
+            return response()->json([
+                'message' => $e->getMessage(),
+                'status' => $e->getCode()
+            ], $e->getCode());
         }
     }
 }
